optimise the ppg calculate

+++ ppg_filter.c
@@ -5,7 +5,7 @@
 #define RING_ADD(a,b,m) ((a)+(b))%(m)
 #define RING_SUB(a,b,m) ((a) - (b) >= 0 )? ((a) - (b)) :((a) - (b) + (m))
 
-typedef __int64 INT64;
+typedef long long INT64;
 
 int FIR_COFFES[FIR_FILTER_LENGTH]=
 {
